+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -201,7 +201,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-24T12:43:22.642Z",
+    "2025-10-25T11:11:21.225Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -241,7 +241,7 @@ export const RetrieveDatasetResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-24T12:43:22.642Z"))
+  updated: z.date().default(() => new Date("2025-10-25T11:11:21.225Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {