+++ packages/orq-rc/RELEASES.md
@@ -4559,3 +4559,13 @@ Based on:
 - [typescript v3.14.0-rc.56] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.56] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.56 - packages/orq-rc
+
+## 2025-10-25 11:13:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.639.3 (2.730.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.57] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.57] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.57 - packages/orq-rc